removed timeFormatter
